Migrated certain functions from interface to io for generlized use
